Instrumentation "boundaries" working:
We only start instrumenting in the agent once we see the main
application class come through. Then, we only instrument those classes
loaded in threads that are either the main application thread or
threads forked by that thread. As always, nothing in oshajava.* is
instrumented.

This does NOT yet take care of those JDK classes that are loaded
before the main application class (e.g. java.util.HashMap). We still
need a solution to remap those appropriately.