skip superclass scanning for instrumented superclasses
This fixes errors on (for example) inner classes that extend their outer class. The
inner class would need to be instrumented before its superclass was loaded. Of
course, there's no realistic scenario where the outer class will not be instrumented
but the inner one will, so checking this prevents errors.